feat(phase-0): land guards+telemetry burn-down wave (#1152 #1153 #1155 #1159 #1178)
Port the phase-0 snapshot onto current main (post #1179-#1188),
reconciling with the shortened artifact locks (#1180), the recoverable
object layout (#1182), the dylint 6.0.1 stack (#1184), and the absorbed
build_dylint_driver logic (#1187):

- Central artifact layout resolver (zccache-artifact/src/layout.rs) +
  persist/layout.rs; runtime consumers no longer format legacy
  <key>_<index> paths (#1152)
- Dylints ban_legacy_artifact_path + ban_normalized_path_deref_containment
  on the 6.0.1/nightly-2026-05-26 stack, wired into ci/lint.py + ci.yml,
  with ci/check_dylint_wiring.py guarding the wiring (#1152 #1178)
- Typed materialization outcomes, concrete journal miss reasons, and
  unknown-miss warnings threaded through connection/compile paths (#1155)
- Rule-based log-audit engine (zccache-audit::log_audit) + shared
  registry ci/log_audit_registry.json + perf/integration gating (#1159)
- Jobserver drop assertion keyed on fstat dev/ino identity (#1153)
- Split 1590-LOC connection.rs into connection/{mod,dispatch,attribution}
  per the file-size convention; tests moved to server/tests/connection_*

Validated: fmt clean; workspace check + clippy under -D warnings; 612
daemon-core unit tests (+ artifact 83, audit 9, core 112, fscache 118);
dylint wiring + log-audit registry checks; pytest ci tests.
