+++ crates/zccache-daemon-core/src/daemon/server/embedded.rs
@@ -223,8 +223,8 @@ impl EmbeddedDaemon {
         // cache_store) attribute their sub-phase records to this compile_id.
         // No-op unless ZCCACHE_INNER_TRACE is set; the IPC wrapper path does
         // not open a scope, so only embedded compiles emit sub-phase records.
-        let (mut response, attributed_miss_reason) = capture_miss_reason(Box::pin(
-            super::inner_trace::scope(
+        let (mut response, attributed_miss_reason) =
+            capture_miss_reason(Box::pin(super::inner_trace::scope(
                 compile_id.clone(),
                 handle_compile_ephemeral(
                     &self.state,
@@ -236,9 +236,8 @@ impl EmbeddedDaemon {
                     request.env,
                     request.stdin,
                 ),
-            ),
-        ))
-        .await;
+            )))
+            .await;
         crate::compile_trace::record(
             "embedded_daemon_compile",
             total.elapsed().as_micros() as u64,
@@ -265,13 +264,7 @@ impl EmbeddedDaemon {
                     latency_ns,
                 );
             }
-            let entry = JournalEntry::new(
-                journal_ctx,
-                outcome,
-                exit_code,
-                latency_ns,
-                miss_reason,
-            );
+            let entry = JournalEntry::new(journal_ctx, outcome, exit_code, latency_ns, miss_reason);
             self.state.journal.log(&entry, None);
         }
         match response {
