+++ crates/zccache-daemon-core/src/daemon/server/persist/mod.rs
@@ -28,8 +28,8 @@ mod artifact_io;
 mod directory_bundle;
 mod fs_caps;
 mod hardlink;
-mod link_registry;
 mod layout;
+mod link_registry;
 mod mtime;
 mod object_store;
 mod pack;
@@ -42,8 +42,8 @@ pub(in crate::daemon::server) use artifact_io::*;
 pub(in crate::daemon::server) use directory_bundle::*;
 pub(in crate::daemon::server) use fs_caps::*;
 pub(in crate::daemon::server) use hardlink::*;
-pub(in crate::daemon::server) use link_registry::*;
 pub(in crate::daemon::server) use layout::*;
+pub(in crate::daemon::server) use link_registry::*;
 pub(in crate::daemon::server) use mtime::*;
 pub(in crate::daemon::server) use object_store::*;
 pub(in crate::daemon::server) use pack::*;
