+++ crates/zccache-daemon-core/src/daemon/server/tests/mod.rs
@@ -10,6 +10,9 @@ mod cache_trim;
 mod clear_handler;
 mod client_env;
 mod compiler_hash;
+mod connection_disconnect;
+mod connection_ipc;
+mod connection_self_profile;
 mod deferred_cold_path;
 mod disk_maintenance;
 mod embedded_flush;
